feat: documentation has been added
