+++ src/event-manager/EventManager.ts
@@ -32,7 +32,7 @@ import type {
 import type {
   GenericSegmentifyEventPayload,
   KEYS_OF_EVENTS,
-} from '../types/interfaces/segmentify-events/events-payload.interface';
+} from '../types/interfaces/segmentify-events/EventsPayloads.interface';
 
 /**
  * RequestCredentials
