+++ src/types/enums/DeviceOs.enum.ts
@@ -1,0 +1,21 @@
+/**
+ * @enum OS_TYPES
+ * @description This enum describes the types of the operating systems.
+ * @property {string} android
+ * @property {string} ios
+ */
+export const OS_TYPES = {
+  android: 'android',
+  ios: 'os',
+} as const;
+
+/**
+ * @enum DEVICE_TYPES
+ * @description This enum describes the types of the devices.
+ * @property {string} android
+ * @property {string} ios
+ */
+export const DEVICE_TYPES = {
+  android: 'android',
+  ios: 'ios',
+} as const;
