+++ src/types/enums/SupportedLanguages.enum.ts
@@ -1,0 +1,16 @@
+/**
+ * Supported languages
+ * @enum SUPPORTED_LANGUAGES
+ * @description Supported languages by Segmentify.
+ * @property {string} TR
+ * @property {string} EN
+ * @property {string} FR
+ * @property {string} DE
+ */
+
+export const SUPPORTED_LANGUAGES = {
+  TR: 'TR',
+  EN: 'EN',
+  FR: 'FR',
+  DE: 'DE',
+} as const;
