--- src/types/enums/api-error.enum.ts
@@ -1,11 +1,0 @@
-export const API_ERROR_ENUMS = {
-  NO_API_KEY: 'NO_API_KEY',
-  NO_ACCOUNT: 'NO_ACCOUNT',
-  UNVERIFIED_ACCOUNT: 'UNVERIFIED_ACCOUNT',
-  FAILED_ACCOUNT: 'FAILED_ACCOUNT',
-  NO_DOMAIN_APIKEY_MATCH: 'NO_DOMAIN_APIKEY_MATCH',
-  NO_EVENT: 'NO_EVENT',
-  NO_USERID: 'NO_USERID',
-  NO_SESSIONID: 'NO_SESSIONID',
-  BAD_INPUT: 'BAD_INPUT',
-} as const;
