--- src/types/enums/device-os.enum.ts
@@ -1,9 +1,0 @@
-export const OS_TYPES = {
-  android: 'android',
-  ios: 'os',
-} as const;
-
-export const DEVICE_TYPES = {
-  android: 'android',
-  ios: 'ios',
-} as const;
