--- src/types/enums/supported-languages.enum.ts
@@ -1,6 +1,0 @@
-export const SUPPORTED_LANGUAGES = {
-  TR: 'TR',
-  EN: 'EN',
-  FR: 'FR',
-  DE: 'DE',
-} as const;
