+++ src/types/index.ts
@@ -29,12 +29,12 @@ export type {
   TRequestPushNotificationPermissionPayload,
   TFirePushNotificationInteraction,
   TNotificationConfigs,
-  PageViewEventType,
-  ProductViewEventType,
-  BasketOperationsEventType,
-  CheckOutEventType,
-  UserOperationsEventType,
-  CustomEventType,
-  InteractionEventType,
-  SearchEventType,
+  PageView,
+  ProductView,
+  BasketOperations,
+  CheckOut,
+  UserOperations,
+  Custom,
+  Interaction,
+  Search,
 } from './interfaces/';
