+++ src/types/interfaces/segmentify-events/BasketOperations.interface.ts
@@ -2,7 +2,21 @@ import type { CommonEventParameters } from '../CommonEventParameters';
 import type { TBASKET_OPERATIONS_STEPS } from '../../types';
 
 /**
- * BasketOperations - This interface describes structure of BasketOperations event's attributes type.
+ * @typedef
+ * @name BasketOperations
+ * @description
+ * BasketOperations is a type that describes the parameters that basket operations event takes.
+ * @property {TBASKET_OPERATIONS_STEPS} step The step of the basket operation. It can be 'add' or 'remove'.
+ * @property {string} productId The id of the product.
+ * @property {number} [price] The price of the product.
+ * @property {number} [quantity] The quantity of the product.
+ * @example
+ *  {
+ *   step: 'add',
+ *   productId: '1234567890',
+ *   price: 250.99,
+ *   quantity: 1,
+ *  }
  */
 export interface BasketOperations extends CommonEventParameters {
   /**
