+++ src/types/interfaces/segmentify-events/index.ts
@@ -4,5 +4,12 @@ export type { BasketOperations } from './BasketOperations.interface';
 export type { CheckOut } from './CheckOut.interface';
 export type { UserOperations } from './UserOperations.interface';
 export type { Custom } from './CustomEvent.interface';
-export type { Interaction } from './interaction.interface';
+export type { Interaction } from './Interaction.interface';
 export type { Search } from './Search.interface';
+export type {
+  TRequestPushNotificationPermission,
+  TFirePushNotification,
+  TRequestPushNotificationPermissionPayload,
+  TFirePushNotificationInteraction,
+  TNotificationConfigs,
+} from './Push.interface';
