+++ src/types/types/BannerDetail.type.ts
@@ -1,0 +1,50 @@
+/**
+ * @typedef
+ * @name BannerDetail
+ * @category Types
+ * @description BannerDetail type is used to define the type of the banner detail.
+ * @property {string} [title] - Banner title
+ * @property {string} [image] - Banner image
+ * @property {number} [order] - Banner order
+ * @property {string[]} [urls] - Banner urls
+ * @type {BannerDetail}
+ * @example
+ * {
+ *  "title": "Banner Title",
+ *  "image": "https://cdn.segmentify.com/demo/banner.jpg",
+ *  "order": 1,
+ *  "urls": [
+ *   "https://www.segmentify.com",
+ *   "https://www.segmentify.com"
+ *  ]
+ * }
+ */
+export type BannerDetail = {
+  /**
+   * title is the title of the banner.
+   * @example
+   * "Banner Title"
+   */
+  title?: string;
+  /**
+   * image is the image of the banner.
+   * @example
+   * "https://cdn.segmentify.com/demo/banner.jpg"
+   */
+  image?: string;
+  /**
+   * order is the order of the banner.
+   * @example
+   * 1
+   */
+  order?: number;
+  /**
+   * urls is the urls of the banner.
+   * @example
+   * [
+   * "https://www.segmentify.com",
+   * "https://www.segmentify.com"
+   * ]
+   */
+  urls?: string[];
+};
