--- src/types/types/banner-detail.type.ts
@@ -1,6 +1,0 @@
-export type BannerDetail = {
-  title?: String;
-  image?: String;
-  order?: Number;
-  urls?: String[];
-};
