--- src/types/types/checkout-product.type.ts
@@ -1,6 +1,0 @@
-export type CheckoutProduct = {
-  productId: String;
-  price?: Number;
-  quantity?: Number;
-  currency?: String;
-};
