--- src/types/types/firebase-push-response.type.ts
@@ -1,16 +1,0 @@
-export type TFireBasePushResponse = {
-  sentTime: number;
-  data: {
-    dataCenterUrl: string;
-    instanceId: string;
-    image: string;
-    title: string;
-    badge: string;
-    sound: string;
-    body: string;
-    apiKey: string;
-  };
-  messageId: string;
-  ttl: number;
-  from: string;
-};
