--- src/types/types/response-code.type.ts
@@ -1,3 +1,0 @@
-import { API_ERROR_ENUMS } from '../enums';
-
-export type ResponseStatusCodeType = keyof typeof API_ERROR_ENUMS | 'SUCCESS';
